convert tests to use pytest
